feat: created AuthController
