+++ app/Policies/DestinationPolicy.php
@@ -8,21 +8,6 @@ use App\Models\User;
 
 class DestinationPolicy
 {
-    /**
-     * Determine whether the user can view any models.
-     */
-    // public function viewAny(User $user): bool
-    // {
-    //     //
-    // }
-
-    // /**
-    //  * Determine whether the user can view the model.
-    //  */
-    // public function view(User $user, Destination $destination): bool
-    // {
-    //     //
-    // }
 
     /**
      * Determine whether the user can create models.
@@ -48,19 +33,4 @@ class DestinationPolicy
         return $user->id === $destination->user_id;
     }
 
-    /**
-     * Determine whether the user can restore the model.
-     */
-    // public function restore(User $user, Destination $destination): bool
-    // {
-    //     //
-    // }
-
-    /**
-     * Determine whether the user can permanently delete the model.
-     */
-    // public function forceDelete(User $user, Destination $destination): bool
-    // {
-    //     //
-    // }
 }
