+++ routes/api.php
@@ -1,13 +1,12 @@
 <?php
 use App\Http\Controllers\Api\DestinationController;
-use App\Http\Controllers\Api\LoginController;
 use App\Http\Controllers\Api\UserController;
-use App\Http\Controllers\Api\RegisterController;
+use App\Http\Controllers\Api\AuthController;
 use Illuminate\Support\Facades\Route;
 
-Route::post('login', [LoginController::class, 'login']);
-Route::post('register', [RegisterController::class, 'register']);
-Route::post('logout', [LoginController::class, 'logout'])->middleware('auth:sanctum');
+Route::post('login', [AuthController::class, 'login']);
+Route::post('register', [AuthController::class, 'register']);
+Route::post('logout', [AuthController::class, 'logout'])->middleware('auth:sanctum');
 
 Route::get('users', [UserController::class, 'index']);
